% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_inputs}
\alias{check_inputs}
\title{Generic input checks}
\usage{
check_inputs(
  api_key = NULL,
  cache_dir = NULL,
  Ticker = NULL,
  SimFinId = NULL,
  statement = NULL,
  period = NULL,
  fyear = NULL,
  start = NULL,
  end = NULL,
  ttm = NULL,
  shares = NULL,
  ratios = NULL
)
}
\arguments{
\item{api_key}{See function using this argument.}

\item{cache_dir}{See function using this argument.}

\item{Ticker}{See function using this argument.}

\item{SimFinId}{See function using this argument.}

\item{statement}{See function using this argument.}

\item{period}{See function using this argument.}

\item{fyear}{See function using this argument.}

\item{start}{See function using this argument.}

\item{end}{See function using this argument.}

\item{ttm}{See function using this argument.}

\item{shares}{See function using this argument.}

\item{ratios}{See function using this argument.}
}
\description{
This function covers all kinds of (recurring) input checks in \link{simfinapi}. This keeps the other functions cleaner.
}
