% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_clean_names}
\alias{set_clean_names}
\title{Clean names}
\usage{
set_clean_names(DT)
}
\arguments{
\item{DT}{A \link{data.table}.}
}
\value{
A \link{data.table} with cleaned names and labels.
}
\description{
Cleans column names so that they comply to R naming conventions.
Keeps the original names in the \code{label} attribute.
}
