% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_find_id.R
\name{sfa_find_id_}
\alias{sfa_find_id_}
\title{Find single SimFin ID's}
\usage{
sfa_find_id_(find, by = "ticker", api_key = Sys.getenv("sfa_api_key"))
}
\arguments{
\item{find}{\verb{[character(1)]} The string to search for.}

\item{by}{\verb{[character(1)]} Either search by "ticker" or by "name".}

\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use \code{options(sfa_api_key = "yourapikey")}.}
}
\description{
Find single SimFin ID's
}
