% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_id.R
\name{sfa_get_id}
\alias{sfa_get_id}
\title{Find one or more SimFin ID's by ticker or name}
\usage{
sfa_get_id(find, by = "ticker", api_key = Sys.getenv("sfa_api_key"))
}
\arguments{
\item{find}{\verb{[character]} The string to search for.}

\item{by}{\verb{[character(1)]} Either search by "ticker" or by "name".}

\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use \code{options(sfa_api_key = "yourapikey")}.}
}
\description{
Find one or more SimFin ID's by ticker or name
}
