% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_id.R
\name{sfa_get_id_}
\alias{sfa_get_id_}
\title{Find single SimFin ID's}
\usage{
sfa_get_id_(find, type, api_key = getOption("sfa_api_key"))
}
\arguments{
\item{find}{\link{character} The string to search for.}

\item{type}{\link{character} Either search by "ticker" (default) or by
"name-search".}

\item{api_key}{\link{character} Your SimFin API key. I recommend setting the
API key globally via \code{options(sfa_api_key = "yourapikey")} and omit the
\code{api_key} argument in all simfinapi functions.}
}
\description{
Find single SimFin ID's
}
