% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_info.R
\name{sfa_get_info}
\alias{sfa_get_info}
\title{Get basic company information}
\usage{
sfa_get_info(
  Ticker = NULL,
  SimFinId = NULL,
  api_key = getOption("sfa_api_key"),
  cache_dir = getOption("sfa_cache_dir")
)
}
\arguments{
\item{Ticker}{\link{integer} Ticker of the companies of interest.}

\item{SimFinId}{\link{integer} 'SimFin' IDs of the companies of interest. Any
SimFinId will be internally translated to the respective \code{Ticker}. This
reduces the number of queries if you would query the same company via
\code{Ticker} \emph{and} \code{SimFinId}.}

\item{api_key}{\link{character} Your 'SimFin' API key. It's recommended to set
the API key globally using \link{sfa_set_api_key}.}

\item{cache_dir}{\link{character} Your cache directory. It's recommended to set
the cache directory globally using \link{sfa_set_cache_dir}.}
}
\description{
Get basic company information
}
