% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_info.R
\name{sfa_get_info_}
\alias{sfa_get_info_}
\title{Get basic company information}
\usage{
sfa_get_info_(ticker, api_key, cache_dir)
}
\arguments{
\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use
\code{options(sfa_api_key = "yourapikey")}.}

\item{simId}{\verb{[integer(1)]} SimFin ID of the company of interest.}
}
\description{
Internal function.
}
