% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_statement.R
\name{sfa_get_pl}
\alias{sfa_get_pl}
\title{Get Profit and Loss Statements}
\usage{
sfa_get_pl(
  simIds,
  period = "TTM",
  fin_year,
  api_key = getOption("sfa_api_key")
)
}
\arguments{
\item{simIds}{\verb{[integer]} SimFin IDs of the companies of interest.}

\item{period}{\verb{[character(1)]} One of "Q1" "Q2" "Q3" "Q4" "H1" "H2" "9M" "FY"
"TTM". See \code{ptype} on
https://simfin.com/api/v1/documentation/#operation/getCompStatementStandardised
for details.}

\item{fin_year}{\verb{[integer(1)]} The financial year of interest.}

\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use
\code{options(sfa_api_key = "yourapikey")}.}
}
\description{
Convenience function which calls \link{sfa_get_statement} with
\code{statement = "pl"}.
}
