% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_prices.R
\name{sfa_get_prices}
\alias{sfa_get_prices}
\title{Get price data}
\usage{
sfa_get_prices(
  Ticker = NULL,
  SimFinId = NULL,
  ratios = NULL,
  start = NULL,
  end = NULL,
  api_key = getOption("sfa_api_key"),
  cache_dir = getOption("sfa_cache_dir")
)
}
\arguments{
\item{Ticker}{\link{character}}

\item{SimFinId}{[]}

\item{ratios}{[]}

\item{start}{[]}

\item{end}{[]}

\item{api_key}{\link{character} Your SimFin API key. It's recommended to set
the API key globally using \link{sfa_set_api_key}.}

\item{cache_dir}{\link{character} Your cache directory. It's recommended to set
the cache directory globally using \link{sfa_set_cache_dir}.}
}
\description{
Get price data
}
