% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_ratios.R
\name{sfa_get_ratios_}
\alias{sfa_get_ratios_}
\title{Get basic company information}
\usage{
sfa_get_ratios_(simId, api_key = getOption("sfa_api_key"))
}
\arguments{
\item{simId}{\verb{[integer(1)]} SimFin ID of the company of interest.}

\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use
\code{option(sfa_api_key = "yourapikey")}.}
}
\description{
Get basic company information
}
