% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_ref.R
\name{sfa_get_ref}
\alias{sfa_get_ref}
\title{Download reference data}
\usage{
sfa_get_ref(ref_data, api_key = getOption("sfa_api_key"))
}
\arguments{
\item{ref_data}{\link{character} Either "industries" or "markets".}

\item{api_key}{\link{character} Your 'SimFin' API key. It's recommended to set
the API key globally using \link{sfa_set_api_key}.}
}
\value{
\link{data.table} containing reference data on "industries" or "markets",
depending on \code{ref_data}.
}
\description{
Download reference data on industries and markets.
}
