% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_load_company_info.R
\name{sfa_load_company_info}
\alias{sfa_load_company_info}
\title{General Company Info}
\usage{
sfa_load_company_info(
  api_key = getOption("sfa_api_key"),
  cache_dir = getOption("sfa_cache_dir")
)
}
\arguments{
\item{api_key}{\link{character} Your 'SimFin' API key. It's recommended to set
the API key globally using \link{sfa_set_api_key}.}

\item{cache_dir}{\link{character} Your cache directory. It's recommended to set
the cache directory globally using \link{sfa_set_cache_dir}.}
}
\value{
\code{data.table::data.table()} containing basic company information
}
\description{
Get Information for one or several companies by SimFin ID or ticker.
}
