% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_load__shares_outstanding.R
\name{sfa_load_weighted_shares_outstanding}
\alias{sfa_load_weighted_shares_outstanding}
\title{Weighted Shares Outstanding}
\usage{
sfa_load_weighted_shares_outstanding(
  id = NULL,
  ticker = NULL,
  fyear = NULL,
  period = NULL,
  start = NULL,
  end = NULL,
  ttm = NULL,
  api_key = getOption("sfa_api_key"),
  cache_dir = getOption("sfa_cache_dir")
)
}
\description{
Weighted Shares Outstanding
}
