% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sfa_memoise_fromJSON}
\alias{sfa_memoise_fromJSON}
\title{memoise jsonlite::fromJSON for SimFin API calls}
\usage{
sfa_memoise_fromJSON(cache_dir = tempdir())
}
\arguments{
\item{cache_dir}{\verb{[character(1)]} Path of the cache directory.}
}
\description{
memoise jsonlite::fromJSON for SimFin API calls
}
