% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_set_sfplus.R
\name{sfa_set_sfplus}
\alias{sfa_set_sfplus}
\title{Specify the type of you SimFin account}
\usage{
sfa_set_sfplus(sfplus = TRUE)
}
\arguments{
\item{sfplus}{\link{logical} Defaults to \code{TRUE} to specify that you have a SimFin+ account.}
}
\description{
If you have a SimFin+ account, it is highly recommended to
specify this globally as it makes specifying the \code{sfplus} argument of other
\verb{sfa_*} functions obsolete.

You don't need this function if you don't have a SimFin+ account.
}
\details{
There is no good reason to use \code{sfa_set_sfplus(FALSE)} as all
functions assume this by default.
}
\examples{
\dontrun{
# Tell simfinapi that you have a SimFin+ account
sfa_set_sfplus()
}
}
