% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfinapi-package.R
\docType{package}
\name{simfinapi-package}
\alias{simfinapi}
\alias{simfinapi-package}
\title{simfinapi: Accessing 'SimFin' Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Through simfinapi, you can intuitively access the 'SimFin' Web-API (\url{https://www.simfin.com/}) to make 'SimFin' data easily available in R. To obtain an 'SimFin' API key (and thus to use this package), you need to register at \url{https://app.simfin.com/login}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthiasgomolka/simfinapi}
  \item Report bugs at \url{https://github.com/matthiasgomolka/simfinapi/issues}
}

}
\author{
\strong{Maintainer}: Matthias Gomolka \email{matthias.gomolka@posteo.de}

}
\keyword{internal}
