% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_doc.R
\name{param_doc}
\alias{param_doc}
\title{Parameter documentation}
\arguments{
\item{api_key}{\link{character} Your 'SimFin' API key. It's recommended to set
the API key globally using \link{sfa_set_api_key}.}

\item{cache_dir}{\link{character} Your cache directory. It's recommended to set
the cache directory globally using \link{sfa_set_cache_dir}.}

\item{sfplus}{\link{logical} Set\code{TRUE} if you have a SimFin+ account. It's
recommended to set \code{sfplus} globally using \link{sfa_set_sfplus}.}

\item{ticker}{\link{integer} Ticker of the companies of interest.}

\item{id}{\link{integer} 'SimFin' IDs of the companies of interest. Any
\code{id} will be internally translated to the respective \code{ticker}. This
reduces the number of queries in case you query the same company via
\code{ticker} \emph{and} \code{id}.}

\item{asreported}{\link{logical} If \code{TRUE}, retrieves the as-reported (not restated) data.}

\item{ttm}{\link{logical} If \code{TRUE}, retrieves trailing twelve month periods.}

\item{start}{\link{Date} Filter for the report dates (reserved for SimFin+ users).
With this filter you can filter the statements by the date on which the
reported period ended ('Report Date'). By specifying a value here, only
statements will be retrieved with report dates ending AFTER the specified
date.}

\item{end}{\link{Date} Filter for the report dates (reserved for SimFin+ users).
With this filter you can filter the statements by the date on which the
reported period ended ('Report Date'). By specifying a value here, only
statements will be retrieved with report dates ending BEFORE the specified
date.}

\item{period}{\link{character} Filter for periods. As a non-SimFin+ user, you have
to provide exactly one period. As SimFin+ user, this filter can be omitted
to retrieve all statements available for the company.
\itemize{
\item \code{'q1'}: First fiscal quarter.
\item \code{'q2'}: Second fiscal quarter.
\item \code{'q3'}: Third fiscal quarter.
\item \code{'q4'}: Fourth fiscal quarter.
\item \code{'fy'}: Full fiscal year.
\item \code{'h1'}: First 6 months of fiscal year.
\item \code{'h2'}: Last 6 months of fiscal year.
\item \code{'9m'}: First nine months of fiscal year.
\item \code{'6m'}: Any fiscal 6 month period (first + second half years; reserved
for SimFin+ users).
\item \code{'quarters'}: All quarters (q1 + q2 + q3 + q4; reserved for SimFin+
users).
}}

\item{fyear}{\link{integer} Filter for fiscal year. As a non-SimFin+ user, you
have to provide exactly one fiscal year. As SimFin+ user, this filter can
be omitted to retrieve all data available for the company.}

\item{ratios}{\link{logical} With \code{TRUE}, you can display some price related
ratios along with the share price data (reserved for SimFin+ users). The
ratios that will be displayed are:
\itemize{
\item Market-Cap
\item Price to Earnings Ratio (quarterly)
\item Price to Earnings Ratio (ttm)
\item Price to Sales Ratio (quarterly)
\item Price to Sales Ratio (ttm)
\item Price to Book Value (ttm)
\item Price to Free Cash Flow (quarterly)
\item Price to Free Cash Flow (ttm)
\item Enterprise Value (ttm)
\item EV/EBITDA (ttm)
\item EV/Sales (ttm)
\item EV/FCF (ttm)
\item Book to Market Value (ttm)
\item Operating Income/EV (ttm).
}}
}
\description{
Parameter documentation
}
\section{Parallel processing}{

This function supports parallel processing via \code{future.apply}. If your
machine has several cores (most have), you can make the API calls in
parallel. To do so, define a \code{future::plan()} before calling the function.
}

