% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_entities.R
\name{sfa_get_entities}
\alias{sfa_get_entities}
\title{Get a table of all available SimFin ID's with ticker and name.}
\usage{
sfa_get_entities(api_key = getOption("sfa_api_key"))
}
\arguments{
\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use
\code{options(sfa_api_key = "yourapikey")}.}
}
\description{
Get a table of all available SimFin ID's with ticker and name.
}
