% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_id.R
\name{sfa_get_id_}
\alias{sfa_get_id_}
\title{Find single SimFin ID's}
\usage{
sfa_get_id_(find, type, api_key = getOption("sfa_api_key"))
}
\arguments{
\item{find}{\verb{[character(1)]} The string to search for.}

\item{api_key}{\verb{[character(1)]} Your SimFin API key. I recommend setting the
API key globally via \code{options(sfa_api_key = "yourapikey")} and omit the
\code{api_key} argument in all simfinapi functions.}

\item{by}{\verb{[character(1)]} Either search by "ticker" (default) or by "name".}
}
\description{
Find single SimFin ID's
}
