% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_info.R
\name{sfa_get_info}
\alias{sfa_get_info}
\title{Get basic company information}
\usage{
sfa_get_info(simIds, api_key = options("sfa_api_key"))
}
\arguments{
\item{simIds}{\verb{[integer]} SimFin IDs of the companies of interest.}

\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use \code{sys.setenv(sfa_api_key = "yourapikey")}.}
}
\description{
Get basic company information
}
