% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_info.R
\name{sfa_get_info_}
\alias{sfa_get_info_}
\title{Get basic company information}
\usage{
sfa_get_info_(ticker, api_key, cache_dir)
}
\arguments{
\item{ticker}{\link{integer} Ticker of the companies of interest.}

\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use
\code{options(sfa_api_key = "yourapikey")}.}

\item{cache_dir}{\link{character} Your cache directory. It's recommended to set
the cache directory globally using \link{sfa_set_cache_dir}.}
}
\description{
Internal function.
}
