% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_info.R
\name{sfa_get_info_}
\alias{sfa_get_info_}
\title{Get basic company information}
\usage{
sfa_get_info_(simId, api_key = Sys.getenv("sfa_api_key"))
}
\arguments{
\item{simId}{\verb{[integer(1)]} SimFin ID of the company of interest.}

\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use \code{Sys.setenv(sfa_api_key = "yourapikey")}.}
}
\description{
Get basic company information
}
