% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_ref.R
\name{sfa_get_ref}
\alias{sfa_get_ref}
\title{Download reference data}
\usage{
sfa_get_ref(ref_data)
}
\arguments{
\item{ref_data}{\link{character} Either "industries" or "markets".}
}
\description{
Download reference data
}
