% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_ref.R
\name{sfa_get_ref}
\alias{sfa_get_ref}
\title{Download reference data}
\usage{
sfa_get_ref(ref_data, api_key = getOption("sfa_api_key"))
}
\arguments{
\item{ref_data}{\link{character} Either "industries" or "markets".}

\item{api_key}{\verb{[character(1)]} Your 'SimFin' API key. For simplicity use
\code{options(sfa_api_key = "yourapikey")}.}
}
\description{
Download reference data
}
