% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa_get_statement.R
\name{sfa_get_statement_}
\alias{sfa_get_statement_}
\title{Get basic company information}
\usage{
sfa_get_statement_(
  simId,
  statement,
  period,
  fin_year,
  api_key = Sys.getenv("sfa_api_key")
)
}
\arguments{
\item{simId}{\verb{[integer(1)]} SimFin ID of the company of interest.}

\item{statement}{\verb{[character(1)]} One of "pl" (Profit and Loss), "bs" (Balance Sheet), "cf" (Cash Flow).}

\item{period}{\verb{[character(1)]} One of "Q1" "Q2" "Q3" "Q4" "H1" "H2" "9M" "FY" "TTM". See \code{ptype} on https://simfin.com/api/v1/documentation/#operation/getCompStatementStandardised for details.}

\item{fin_year}{\verb{[integer(1)]} The financial year of interest.}

\item{api_key}{\verb{[character(1)]} Your SimFin API key. For simplicity use \code{Sys.setenv(sfa_api_key = "yourapikey")}.}
}
\description{
Get basic company information
}
