% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares_outstanding.R
\name{sfa_load_common_shares_outstanding}
\alias{sfa_load_common_shares_outstanding}
\title{Common Shares Outstanding}
\usage{
sfa_load_common_shares_outstanding(
  id = NULL,
  ticker = NULL,
  start = NULL,
  end = NULL
)
}
\description{
Common Shares Outstanding
}
