% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares_outstanding.R
\name{sfa_load_weighted_shares_outstanding}
\alias{sfa_load_weighted_shares_outstanding}
\title{Weighted Shares Outstanding}
\usage{
sfa_load_weighted_shares_outstanding(
  id = NULL,
  ticker = NULL,
  start = NULL,
  end = NULL
)
}
\description{
Weighted Shares Outstanding
}
